requirements.txt
requests
feedparser
PyYAML
python-dateutil
tweepy
python-dotenv
